pyproject.toml
[tool.poetry]
name = "wikimpacts"
version = "0.1.0"
description = "Dependencies for parsing and normalizing environmental impact data"
authors = ["shorouq.zahra@ri.se"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
normalizer = "^0.2.1"
dateparser = "^1.2.0"
text2num = "^2.5.0"
pandas = "^2.0.0"
shortuuid = "^1.0.13"
pre-commit = "^3.2.0"
spacy = "^3.7.4"
regex = "^2023.12.25"
num2words = "^0.5.13"
geopy = "^2.4.1"
requests-cache = "^1.2.0"
pycountry = "22.3.5"
fastparquet = "2023.10.1"
wikipedia-api = "^0.6.0"
unidecode = "^1.3.8"
beautifulsoup4 = "^4.12.3"
transformers = "^4.41.2"
torch = "^2.3.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"
openpyxl = "^3.1.2"
numpy = "1.24"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
